Introduce a `Types` class to help with future typecasting.
The nature of the plugin is that it may be in environments with less than great code, it may not return the expected types for example.

This introduces the `Types` class, which aims to provide tools for the Troubleshooting plugin to ensure expected types are used within itself, and in the future, also alert of plugins or themes that may be incorrectly returning types for what the troubleshooting plugin does.

This also removes the typecast from the Troubleshoot class, so that it does not cause fatal errors on activation where other plugins may not be appropriately modifying the plugins page.